cambios capas

+++ src/dao/productManagerMONGO.js
@@ -1,6 +1,11 @@
 
+
 import { productModelo } from "./models/productModelo.js";
 import { userModelo } from "./models/userModelo.js";
+import { logger } from "../utils/logger.js";
+import { CustomError } from "../utils/CustomError.js"
+import { TIPOS_ERRORS } from "../utils/Errors.js"
+
 
 class ProductManagerMONGO {
     async getProducts() {
@@ -21,46 +26,74 @@ class ProductManagerMONGO {
         return await productModelo.paginate({}, options) //2 argumentos: filtro
     }//DEVUELVE UN OBJETO CON PROPIEDADES DOCS (ARRAY CON DOCUMENTOS), TOTAL DOC, LIMITE, ....
     async createProduct({ owner, title, category, description, price, thumbnail, stock, status }) {
-        let userEmail = "admin";
         if (owner) {
             let user = await userModelo.findOne({ email: owner });
             if (user) {
-                userEmail = user.email;
+                owner = user.email;
+            }
+            else {
+                owner = "admin"
             }
         }
-        return await productModelo.create({ owner: userEmail, title, category, description, price, thumbnail, stock, status })
+        else {
+            owner = "admin"
+        }
+        return await productModelo.create({ owner, title, category, description, price, thumbnail, stock, status })
     }
     async getProductsBy(filter) {
         return await productModelo.find(filter).lean()
     }
     async getProductById(id) {
-        return await productModelo.findById(id)
+        return await productModelo.findById(id).lean()
     }
-    async updateProduct(email, id, price) {
+    async updateProduct(id, email, price) {
         let user = await userModelo.findOne({ email: email });
+        if (!user) {
+            logger.warn(`Usuario con email ${email} no encontrado`)
+            return CustomError.createError(
+                "Error al actualizar el producto", `Usuario con email ${email} no encontrado`, TIPOS_ERRORS.NOT_FOUND
+            );
+        }
         let product = await this.getProductById(id)
+        if (!product) {
+            logger.warn(`Producto con ${id} no encintrado`)
+            return CustomError.createError(
+                "Error al actualizar el producto", `Producto con ID ${id} no encontrado`, TIPOS_ERRORS.NOT_FOUND
+            );
+        }
         if (product.owner === user.email) {
             product.price = price
-            console.log("SOLO EL OWNER PUEDE ELIMIAR EL PRODUCOT");
-            product.save();
-            console.log(`NUEVO PRECIO : ${product.price}`);
-            return await productModelo.updateOne({ _id: id }, { price: price })
+            product.save
+            logger.info(`Precio acualizado:${product.title} ${product.price}`)
+            return product
         }
         else {
-            console.log(`privilegios insuficientes para modifica rel producto ${id}: solo usuarios premium o propietarios`);
+            logger.warn(`privilegios insuficientes para modifica el producto ${id}: solo usuarios premium o propietarios`)
         }
     }
     async deleteProduct(email, id) {
         let user = await userModelo.findOne({ email: email });
-        let product = await this.getProductById(id)
+        if (!user) {
+            logger.warn(`Usuario con email ${email} no encontrado`)
+            return CustomError.createError(
+                "Error al actualizar el producto", `Usuario con email ${email} no encontrado`, TIPOS_ERRORS.NOT_FOUND
+            );
+        }
+        let product = await this.getProductById(id).lean()
+        if (!product) {
+            logger.warn(`Producto con ${id} no encintrado`)
+            return CustomError.createError(
+                "Error al actualizar el producto", `Producto con ID ${id} no encontrado`, TIPOS_ERRORS.NOT_FOUND
+            );
+        }
         if (user.rol === "admin" || product.owner === user.email) {
-            console.log(`producto eliminado con exito: ${product.title}`);
+            logger.info(`producto eliminado con exito: ${product.title}`)
             return await productModelo.deleteOne({ _id: id });
-
         } else {
-            console.log(`Privilegios insuficientes para eliminar el producto ${id}: solo usuarios premium o propietarios`);
+            logger.warn(`privilegios insuficientes para modifica el producto ${id}: solo usuarios premium o propietarios`)
         }
     }
 }
+
 
 export { ProductManagerMONGO }
